refactor(phase2): split the AdminPage god-component (A4)
Extract the inlined sub-components from the ~570-line AdminPage into focused
files under ./components: ActionButtons, AdminStateViews (StatCard + Loading/
Empty/Error), OrganizationsList, LocationsList (+ LocationRow), EmployeesList
(+ AdminBadgeWithTooltip). AdminPage now only orchestrates tabs/search/modals
(~290 lines). No behaviour change.
